Implemented BIP0353
